Fix empty enum serialisation
